podcastTextGenerationApp/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.10
    # via langchain
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.40.0
    # via langchain-anthropic
anyio==4.7.0
    # via
    #   anthropic
    #   httpx
    #   openai
astroid==3.3.5
    # via pylint
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==24.2.0
    # via
    #   aiohttp
    #   cattrs
    #   requests-cache
cattrs==24.1.2
    # via requests-cache
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
colorama==0.4.6
    # via -r requirements.in
defusedxml==0.7.1
    # via langchain-anthropic
dill==0.3.9
    # via pylint
distro==1.9.0
    # via
    #   anthropic
    #   openai
exceptiongroup==1.2.2
    # via
    #   anyio
    #   cattrs
feedparser==6.0.11
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   langsmith
    #   openai
    #   tavily-python
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
isort==5.13.2
    # via pylint
jiter==0.8.0
    # via
    #   anthropic
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.10
    # via -r requirements.in
langchain-anthropic==0.2.4
    # via -r requirements.in
langchain-core==0.3.22
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-anthropic
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.2.11
    # via -r requirements.in
langchain-text-splitters==0.3.2
    # via langchain
langsmith==0.1.147
    # via
    #   langchain
    #   langchain-core
mccabe==0.7.0
    # via pylint
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via langchain
openai==1.57.0
    # via langchain-openai
orjson==3.10.12
    # via langsmith
packaging==24.2
    # via langchain-core
platformdirs==4.3.6
    # via
    #   pylint
    #   requests-cache
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.10.3
    # via
    #   anthropic
    #   langchain
    #   langchain-anthropic
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.27.1
    # via pydantic
pylint==3.3.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   -r requirements.in
    #   langchain
    #   langsmith
    #   requests-cache
    #   requests-toolbelt
    #   tavily-python
    #   tiktoken
requests-cache==1.2.1
    # via -r requirements.in
requests-toolbelt==1.0.0
    # via langsmith
sgmllib3k==1.0.0
    # via feedparser
six==1.17.0
    # via
    #   python-dateutil
    #   url-normalize
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
sqlalchemy==2.0.36
    # via langchain
tavily-python==0.5.0
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
tiktoken==0.8.0
    # via
    #   langchain-openai
    #   tavily-python
tinytag==2.0.0
    # via -r requirements.in
tomli==2.2.1
    # via pylint
tomlkit==0.13.2
    # via pylint
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   astroid
    #   cattrs
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   pylint
    #   sqlalchemy
url-normalize==1.4.3
    # via requests-cache
urllib3==2.2.3
    # via
    #   -r requirements.in
    #   requests
    #   requests-cache
yarl==1.18.3
    # via aiohttp
